update flask app - include JSON file of features with defaults
Updated app.py to use the HTML template 'feature_input.html' for a POST request with feature values.  Included 'features.json' which will contain the default values.  In app.py, we will load in the JSON file to check against input values and update accordingly (for dropdown values).  Clicking 'Submit' on the flask app page takes you to the '/price' route which displays the predicted price.